ProjectOne
Celsius to Fahrenheit

Write a program that converts Celsius temperatures to Fahrenheit temperatures. The formula is

Formula for Celsius to Fahrenheit where F is the Fahrenheit temperature and C is the Celsius temperature. Write a program that prompts the user to input a Celsius temperature and then displays the corresponding Fahrenheit temperature.