Lesson 1
Study and create the Chapter 2 Tying It All Together: Smile! program in Visual Studio 2015. Do NOT simply copy and paste the code from a digital source. You can only learn proper coding by TYPING it out. That way, you see each line of code and it can be placed in your long-term memory. Remember, the objective is to LEARN how to program, not simply complete the exercises as fast as possible.
Assignment 3 - Question 3

An electronics company makes circuit boards that cost $14.95 apiece to produce. Write a program to determine how much the company should sell them for if it wants to make a 35 percent profit. Display the result on the screen.