Lesson 2
Study and create the Chapter 3 Case Study-Part 1 program in Visual Studio 2015. Do NOT simply copy and paste the code from a digital source. You can only learn proper coding by TYPING it out. That way, you see each line of code and it can be placed in your long-term memory. Remember, the objective is to LEARN how to program, not simply complete the exercises as fast as possible.
Assignment 3 - Question 2

Study and create the Chapter 3 "Tying it All Together: Word Game" program in Visual Studio 2015. Do NOT simply copy and paste the code from a digital source. You can only learn proper coding by TYPING it out. That way, you see each line of code and it can be placed in your long-term memory. Remember, the objective is to LEARN how to program, not simply complete the exercises as fast as possible.
Assignmnet Three - Question 3

Many financial experts advise property owners to insure their homes or buildings for at least 80 percent of the amount it would cost to replace the structure. Write a program that asks the user to enter the replacement cost of a building and then displays the minimum amount of insurance that should be purchased for the property.
Assignment three - Question 4


Write a program that calculates the average monthly rainfall for three months. The program should ask the user to enter the name of each month, such as June or July, and the amount of rain (in inches) that fell that month. The program should display a message similar to the following:

The average monthly rainfall for June, July, and August was 6.72 inches.